third_party/blink/web_tests/external/wpt/css/css-color/system-color-consistency.html

<!DOCTYPE html>
<html>

<head>
  <meta charset="utf-8">
  <title>CSS Color 4: Computed value of color-contrast()</title>
  <link rel="author" title="Jan Keitel" href="mailto:jkeitel@google.com">
  <link rel="help" href="https://www.w3.org/TR/css-color-4/#css-system-colors">
  <meta name="assert" content="computed style of form elements is consistent with definition of system colors">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
</head>

<body>
  <div id="target"></div>
  <!-- Reference elements -->
  <!-- Buttons -->
  <div id="buttons">
    <button name="button"></button><input type="submit" name="submit button"><input type="reset" name="reset button">
  </div>
  <script>
    for (let element of document.getElementById("buttons").children) {
      style = document.defaultView.getComputedStyle(element);
      test_computed_value(`color`, `ButtonText`, style.getPropertyValue('color'), 'has the same color as text on a ' + element.name);
    }
    // Test with no specified target contrast
  </script>
</body>

</html>
